refactor to grab script, and fix mouse crusor not locking
