+++ movieTheaterSim/Assets/Scripts/ObjectGrabbable.cs
@@ -4,25 +4,12 @@ public class ObjectGrabbable : MonoBehaviour
 {
     private GameObject player;
     private Rigidbody objectRigidbody;
-    private Transform objectGrabPointTransform;
+    public Transform objectGrabPointTransform;
     [SerializeField] private float throwForce = 500f; //force at which the object is thrown at
-
-    [Header("Supply object configurations")]
-    [SerializeField] private bool isSupplies; //Set trun if grabbed object is able to supple items
-    [SerializeField] private float supplyDistance = 1f; //The distance the grabbed object needs to to to supply items
-    private ItemObjectHandler thisItemObjectHandler; //link of the boxItems object child component
-    LayerMask layerMask; //layer needed on the object that is able to take supplys
-    private ItemObjectHandler hitItemObjectHandler; // will be set by raycast, the object able to be supplied
-
-    private void Awake()
-    {
-        objectRigidbody = GetComponent<Rigidbody>();
-        layerMask = LayerMask.GetMask("Suppliable");
-    }
 
     private void Start()
     {
-        this.thisItemObjectHandler = GetComponentInChildren<ItemObjectHandler>();
+        objectRigidbody = GetComponent<Rigidbody>();
     }
     public void Grab(Transform objectGrabPointTransform, GameObject player)
     {
@@ -66,39 +53,5 @@ public class ObjectGrabbable : MonoBehaviour
             objectRigidbody.MovePosition(newPosition);
             transform.rotation = Quaternion.Lerp(transform.rotation,objectGrabPointTransform.rotation, Time.deltaTime * lerpSpeed);
         } 
-    }
-
-    private void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Mouse0))
-        {
-            //checks if player is holding an item and the item is able to supply objects
-            if (objectGrabPointTransform != null && isSupplies)
-            {
-                //cast ray to see if you hit shelf and then if box has item try and add to self
-                if (Physics.Raycast(transform.position, transform.TransformDirection(Vector3.forward), out RaycastHit raycastHit, supplyDistance, layerMask))
-                {
-                    Debug.DrawRay(transform.position, transform.TransformDirection(Vector3.forward) * raycastHit.distance, Color.green, 3f);
-                    Debug.Log("Did Hit");
-                    if (raycastHit.transform.TryGetComponent(out hitItemObjectHandler))
-                    {
-                        //need if item can be removed from grabbable supplier and item can be added
-                        bool canAddShelfItem = hitItemObjectHandler.AbleToAddItem();
-                        bool canRemoveBoxItem = thisItemObjectHandler.AbleToRemoveItem();
-
-                        if (canAddShelfItem && canRemoveBoxItem) 
-                        {
-                            hitItemObjectHandler.AddItem(thisItemObjectHandler.RemoveItem());
-                            this.hitItemObjectHandler = null;
-                        }
-                    }
-                }
-                else
-                {
-                    Debug.DrawRay(transform.position, transform.TransformDirection(Vector3.forward) * 1000, Color.red, 3f);
-                    Debug.Log("Did not Hit");
-                }
-            }
-        }
     }
 }
