+++ movieTheaterSim/Assets/Scripts/SodaMachine.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76ed1d87f41359e4aa2adf3f6773c169
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
