+++ movieTheaterSim/Assets/Scripts/SodaMachine/SodaMachineHandler.cs
@@ -1,0 +1,28 @@
+using UnityEngine;
+
+public class SodaMachineHandler : MonoBehaviour
+{
+    // Start is called once before the first execution of Update after the MonoBehaviour is created
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+
+
+    //TODO
+    //buttons fill soda on/off
+    //soda has fuel
+    //soda machine has a slot for bottle
+
+    //soda bottle is a grabbable
+    //soda has a fill level
+    
+
+}
