+++ movieTheaterSim/README.md
@@ -4,6 +4,6 @@
 keys can be changed and adjusted in engine by going in to: Edit -> Project Settings... -> Input Manager
 
 **Current keys**<br />
-PickUp = e<br />
+Supply= mouse 0<br />
 Throw = mouse 1<br />
-Interaction = i<br />
+Interaction = e<br />
